+++ internal/controller/ledger/store.go
@@ -41,9 +41,9 @@ type Store interface {
 	//  * the reverted transaction
 	//  * a boolean indicating if the transaction has been reverted. false indicates an already reverted transaction (unless error != nil)
 	//  * an error
-	RevertTransaction(ctx context.Context, id int, at time.Time) (*ledger.Transaction, bool, error)
-	UpdateTransactionMetadata(ctx context.Context, transactionID int, m metadata.Metadata, at time.Time) (*ledger.Transaction, bool, error)
-	DeleteTransactionMetadata(ctx context.Context, transactionID int, key string, at time.Time) (*ledger.Transaction, bool, error)
+	RevertTransaction(ctx context.Context, id uint64, at time.Time) (*ledger.Transaction, bool, error)
+	UpdateTransactionMetadata(ctx context.Context, transactionID uint64, m metadata.Metadata, at time.Time) (*ledger.Transaction, bool, error)
+	DeleteTransactionMetadata(ctx context.Context, transactionID uint64, key string, at time.Time) (*ledger.Transaction, bool, error)
 	UpdateAccountsMetadata(ctx context.Context, m map[string]metadata.Metadata) error
 	// UpsertAccount returns a boolean indicating if the account was upserted
 	UpsertAccounts(ctx context.Context, accounts ...*ledger.Account) error
